cleanup benchmarks
