+++ crates/application/benches/compute.rs
@@ -1,3 +1,311 @@
+use commonware_cryptography::{Hasher as _, Sha256, Signer as _, ed25519};
+use commonware_glue::stateful::db::{DatabaseSet, Unmerkleized as _};
+use commonware_parallel::{Rayon, Strategy as _};
+use commonware_runtime::{Runner as _, buffer::paged::CacheRef, deterministic};
+use commonware_storage::{
+    journal::contiguous::fixed::Config as FixedJournalConfig, merkle::full::Config as MmrConfig,
+    qmdb::any::FixedConfig, translator::EightCap,
+};
+use commonware_utils::{NZU16, NZU64, NZUsize};
+use constantinople_application::{
+    consensus::{self, StateBatch, StateDatabase},
+    executor::PreparedTransfer,
+};
+use constantinople_primitives::{
+    Account, AccountKey, Nonce, Transaction, TransactionPublicKey, VerifiedTransaction,
+};
+use core::num::NonZeroU64;
+use std::{
+    hint::black_box,
+    time::{Duration, Instant},
+};
+
+type Db = StateDatabase<deterministic::Context, Sha256, EightCap, Rayon>;
+type Batch = StateBatch<deterministic::Context, Sha256, EightCap, Rayon>;
+type TestTx = VerifiedTransaction<Sha256>;
+
+const ACCOUNTS: u64 = 1_000_000;
+const COUNTS: &[usize] = &[16_384, 32_768];
+const MAX_SIGNED_ACCOUNTS: u64 = 65_536;
+const NAMESPACE: &[u8] = b"compute-bench";
+const SHARED_FANOUT: usize = 8;
+const WARMUP: u32 = 2;
+const ITERS: u32 = 10;
+
+#[derive(Clone, Copy)]
+enum Fixture {
+    Unique,
+    Shared,
+}
+
+impl Fixture {
+    const fn name(self) -> &'static str {
+        match self {
+            Self::Unique => "unique",
+            Self::Shared => "shared",
+        }
+    }
+}
+
+fn key(index: u64) -> AccountKey {
+    AccountKey::try_from(Sha256::hash(&index.to_le_bytes()).as_ref()).expect("32-byte key")
+}
+
+fn signed_key(index: u64) -> AccountKey {
+    AccountKey::from_public_key(&TransactionPublicKey::ed25519(
+        ed25519::PrivateKey::from_seed(index).public_key(),
+    ))
+}
+
+struct TestSigner {
+    key: ed25519::PrivateKey,
+    public_key: ed25519::PublicKey,
+}
+
+impl TestSigner {
+    fn from_seed(seed: u64) -> Self {
+        let key = ed25519::PrivateKey::from_seed(seed);
+        let public_key = key.public_key();
+        Self { key, public_key }
+    }
+
+    fn sign(&self, to: ed25519::PublicKey, value: u64, nonce: u64) -> TestTx {
+        Transaction::new(
+            TransactionPublicKey::ed25519(self.key.public_key()),
+            TransactionPublicKey::ed25519(to),
+            NonZeroU64::new(value).expect("bench value must be non-zero"),
+            nonce,
+        )
+        .seal_and_sign(&self.key, NAMESPACE, &mut Sha256::default())
+    }
+}
+
+fn config(strategy: Rayon, cache: CacheRef) -> FixedConfig<EightCap, Rayon> {
+    FixedConfig {
+        merkle_config: MmrConfig {
+            journal_partition: "bench-state-journal".into(),
+            metadata_partition: "bench-state-metadata".into(),
+            items_per_blob: NZU64!(1 << 20),
+            write_buffer: NZUsize!(1 << 20),
+            strategy,
+            page_cache: cache.clone(),
+        },
+        journal_config: FixedJournalConfig {
+            partition: "bench-state-log".into(),
+            items_per_blob: NZU64!(1 << 20),
+            page_cache: cache,
+            write_buffer: NZUsize!(1 << 20),
+        },
+        translator: EightCap,
+    }
+}
+
+fn transfers(fixture: Fixture, n: usize) -> Vec<PreparedTransfer> {
+    match fixture {
+        Fixture::Unique => (0..n)
+            .map(|i| {
+                let sender = key(i as u64);
+                let recipient = key(n as u64 + i as u64);
+                PreparedTransfer {
+                    sender,
+                    recipient,
+                    sender_prefix: sender.prefix(),
+                    recipient_prefix: recipient.prefix(),
+                    value: 1,
+                    nonce: 0,
+                }
+            })
+            .collect(),
+        Fixture::Shared => {
+            let accounts = (n / SHARED_FANOUT).max(1);
+            let mut nonces = vec![0u64; accounts];
+            (0..n)
+                .map(|i| {
+                    let sender_index = i % accounts;
+                    let recipient_index = (i * 7 + 3) % accounts;
+                    let nonce = nonces[sender_index];
+                    nonces[sender_index] += 1;
+                    let sender = key(sender_index as u64);
+                    let recipient = key(recipient_index as u64);
+                    PreparedTransfer {
+                        sender,
+                        recipient,
+                        sender_prefix: sender.prefix(),
+                        recipient_prefix: recipient.prefix(),
+                        value: 1,
+                        nonce,
+                    }
+                })
+                .collect()
+        }
+    }
+}
+
+fn signed_txs(fixture: Fixture, n: usize) -> Vec<TestTx> {
+    match fixture {
+        Fixture::Unique => (0..n)
+            .map(|i| {
+                let sender = TestSigner::from_seed(i as u64);
+                let recipient = TestSigner::from_seed(n as u64 + i as u64).public_key;
+                sender.sign(recipient, 1, 0)
+            })
+            .collect(),
+        Fixture::Shared => {
+            let accounts = (n / SHARED_FANOUT).max(1);
+            let signers = (0..accounts)
+                .map(|index| TestSigner::from_seed(index as u64))
+                .collect::<Vec<_>>();
+            let mut nonces = vec![0u64; accounts];
+            (0..n)
+                .map(|i| {
+                    let sender_index = i % accounts;
+                    let recipient_index = (i * 7 + 3) % accounts;
+                    let nonce = nonces[sender_index];
+                    nonces[sender_index] += 1;
+                    signers[sender_index].sign(
+                        signers[recipient_index].public_key.clone(),
+                        1,
+                        nonce,
+                    )
+                })
+                .collect()
+        }
+    }
+}
+
+async fn time_compute(
+    batch: &Batch,
+    strategy: &Rayon,
+    transfers: &[PreparedTransfer],
+) -> (usize, Duration) {
+    let start = Instant::now();
+    let writes = consensus::compute(batch, strategy, transfers)
+        .await
+        .expect("compute path");
+    let elapsed = start.elapsed();
+    let count = writes.len();
+    black_box(&writes);
+    (count, elapsed)
+}
+
+async fn time_prepare_compute(
+    batch: &Batch,
+    strategy: &Rayon,
+    txs: &[TestTx],
+) -> (usize, Duration) {
+    let start = Instant::now();
+    let (transfers, digests) = consensus::prepare_signed(strategy, txs).expect("prepare");
+    let writes = consensus::compute(batch, strategy, &transfers)
+        .await
+        .expect("compute path");
+    let elapsed = start.elapsed();
+    let count = writes.len();
+    black_box((&transfers, &digests, &writes));
+    (count, elapsed)
+}
+
+fn env_u32(name: &str, default: u32) -> u32 {
+    std::env::var(name)
+        .ok()
+        .and_then(|value| value.parse::<u32>().ok())
+        .unwrap_or(default)
+}
+
 fn main() {
-    constantinople_application::consensus::bench::compute();
+    deterministic::Runner::default().start(|context| async move {
+        let bench_prepare = std::env::var_os("CONSTANTINOPLE_BENCH_PREPARE").is_some();
+        let warmup = env_u32("CONSTANTINOPLE_BENCH_WARMUP", WARMUP);
+        let iters = env_u32("CONSTANTINOPLE_BENCH_ITERS", ITERS).max(1);
+        let strategy = Rayon::new(NZUsize!(8)).expect("rayon pool");
+        let cache = CacheRef::from_pooler(&context, NZU16!(8192), NZUsize!(65536));
+        let db = <Db as DatabaseSet<deterministic::Context>>::init(
+            context,
+            config(strategy.clone(), cache),
+        )
+        .await;
+
+        let mut batch = db.new_batches().await;
+        for index in 0..ACCOUNTS {
+            batch = batch.write(
+                key(index),
+                Some(Account {
+                    balance: 1_000_000,
+                    nonce: Nonce::default(),
+                }),
+            );
+        }
+        if bench_prepare {
+            for index in 0..MAX_SIGNED_ACCOUNTS {
+                batch = batch.write(
+                    signed_key(index),
+                    Some(Account {
+                        balance: 1_000_000,
+                        nonce: Nonce::default(),
+                    }),
+                );
+            }
+        }
+        let merkleized = batch.merkleize().await.expect("seed merkleize");
+        db.finalize(merkleized).await;
+
+        let fixture_filter = std::env::var("CONSTANTINOPLE_BENCH_FIXTURE").ok();
+        let count_filter = std::env::var("CONSTANTINOPLE_BENCH_COUNT")
+            .ok()
+            .and_then(|count| count.parse::<usize>().ok());
+        for &n in COUNTS {
+            if count_filter.is_some_and(|filter| filter != n) {
+                continue;
+            }
+            for fixture in [Fixture::Unique, Fixture::Shared] {
+                if fixture_filter
+                    .as_deref()
+                    .is_some_and(|filter| filter != fixture.name())
+                {
+                    continue;
+                }
+
+                let transfers = transfers(fixture, n);
+                let mut total = Duration::ZERO;
+                let mut writes = 0usize;
+                for iter in 0..(warmup + iters) {
+                    let batch = db.new_batches().await;
+                    let (count, elapsed) = time_compute(&batch, &strategy, &transfers).await;
+                    writes = count;
+                    if iter >= warmup {
+                        total += elapsed;
+                    }
+                }
+
+                let avg = total / iters;
+                let tps = n as f64 / avg.as_secs_f64() / 1e6;
+                println!(
+                    "compute  {n} txs / {ACCOUNTS} accounts / {} / {} shards\n  compute: {avg:?}  ({tps:.2} Melem/s) / {writes} writes",
+                    fixture.name(),
+                    strategy.parallelism_hint().max(1),
+                );
+
+                if bench_prepare {
+                    let txs = signed_txs(fixture, n);
+                    let mut total = Duration::ZERO;
+                    let mut writes = 0usize;
+                    for iter in 0..(warmup + iters) {
+                        let batch = db.new_batches().await;
+                        let (count, elapsed) = time_prepare_compute(&batch, &strategy, &txs).await;
+                        writes = count;
+                        if iter >= warmup {
+                            total += elapsed;
+                        }
+                    }
+
+                    let avg = total / iters;
+                    let tps = n as f64 / avg.as_secs_f64() / 1e6;
+                    println!(
+                        "prepare+compute  {n} txs / {ACCOUNTS} accounts / {} / {} shards\n  compute: {avg:?}  ({tps:.2} Melem/s) / {writes} writes",
+                        fixture.name(),
+                        strategy.parallelism_hint().max(1),
+                    );
+                }
+            }
+        }
+    });
 }
