--- crates/application/benches/lazy_body_prepare.rs
@@ -1,3 +1,0 @@
-fn main() {
-    constantinople_application::consensus::bench::lazy_body_prepare();
-}
